new tutorials scripts which cover all of our website tutorials. These are verified nightly on our test machines

+++ Tutorials/qmechanicTutorials.sh
@@ -1,17 +1,27 @@
 #!/bin/bash
-set -e
+#set -e
+
+if [ -z "${QBHOME}" ]; then
+    echo "ERROR: QBHOME is not set! Must set the QBHOME environment variable THEN call this function"
+    exit
+fi
+
+DIVCON_BIN=${QBHOME}/bin/qmechanic
+
+if [ ! -x "${DIVCON_BIN}" ]; then
+    echo "ERROR: ${DIVCON_BIN} is not an executable! Set this path to the qmechanic executable"
+    exit
+fi
 
 currentDate=`date`
-echo "BEGIN Tutorial Test at ${currentDate}"
+echo "BEGIN Tutorial Test at ${currentDate} using ${DIVCON_BIN}"
 
-#UPDATE DIVCON_BIN to the proper path to the qmechanic binary:
-DIVCON_BIN=/home/lance/Release/DivConDiscoverySuite-DEV.895/bin/qmechanic
 WORKDIR=$PWD
 
 echo "First Tutorial: Single Point Calculation"
 rm -rf ${WORKDIR}/SinglePointCalculation ; mkdir -p ${WORKDIR}/SinglePointCalculation ; cd ${WORKDIR}/SinglePointCalculation
 
-${DIVCON_BIN} 4EK4 --prepare --np 8 -v 2 -p pdb 
+wget http://downloads.quantumbioinc.com/media/tutorials/cli/4EK4_out.pdb
 mv 4EK4_out.pdb 4EK4.pdb
 ${DIVCON_BIN} 4EK4.pdb --np 8 -h pm6 -v 2 -p pdb 
 
@@ -25,7 +35,7 @@ echo "Third Tutorial: Structure Optimization"
 rm -rf ${WORKDIR}/StructureOptimization ; mkdir -p ${WORKDIR}/StructureOptimization ; cd ${WORKDIR}/StructureOptimization
 
 wget http://downloads.quantumbioinc.com/media/tutorials/cli/1LRI-addH.pdb
-${DIVCON_BIN} 1LRI-addH.pdb -h pm6 --np 8 -v 2 --opt all 5 0.01 --symmetry -p pdb
+${DIVCON_BIN} 1LRI-addH.pdb -h pm6 --np 8 -v 2 --opt all 3 0.01 --symmetry -p pdb
 
 echo "Forth Tutorial: Active Site Structure Optimization"
 rm -rf ${WORKDIR}/ActiveSiteStructureOptimization ; mkdir -p ${WORKDIR}/ActiveSiteStructureOptimization ; cd ${WORKDIR}/ActiveSiteStructureOptimization
@@ -36,7 +46,8 @@ ${DIVCON_BIN} 1P2Y-addH.pdb --opt /*/HEM/*//,/*/NCT/*// 5.0 0.0 --np 8 -h pm6 -v
 echo "Fifth Tutorial: ONIOM (mixed-QM/MM) Simulations"
 rm -rf ${WORKDIR}/ONIOM_Simulations ; mkdir -p ${WORKDIR}/ONIOM_Simulations ; cd ${WORKDIR}/ONIOM_Simulations
 
-${DIVCON_BIN} 1LRI --prepare -h pm6 amberff14 --opt 100 0.01 --region-selection /A/CLR/99// 3.0 0.0 --np 8 -v 2 -p pdb
+wget http://downloads.quantumbioinc.com/media/tutorials/cli/1LRI-addH.pdb
+${DIVCON_BIN} 1LRI-addH.pdb -h pm6 amberff14sb --opt 100 0.01 --qm-region /A/CLR/99// 3.0 0.0 --np 8 -v 2 -p pdb
 
 echo "Sixth Tutorial: Protonation"
 rm -rf ${WORKDIR}/Protonation ; mkdir -p ${WORKDIR}/Protonation ; cd ${WORKDIR}/Protonation
@@ -44,4 +55,4 @@ rm -rf ${WORKDIR}/Protonation ; mkdir -p ${WORKDIR}/Protonation ; cd ${WORKDIR}/
 ${DIVCON_BIN} 4EK4 --prepare --np 8 -v 2 -p pdb
 
 currentDate=`date`
-echo "END Tutorial Test at ${currentDate}"
+echo "END Tutorial Test at ${currentDate} using ${DIVCON_BIN}"
